cstracker now works from one script
